Isolate search2 types
